Several improvements to the motion-curve-editor

+++ src/org/jwildfire/create/tina/variation/iflames/VectorMath.java
@@ -19,20 +19,20 @@ package org.jwildfire.create.tina.variation.iflames;
 import org.jwildfire.base.mathlib.MathLib;
 
 public class VectorMath {
-  public static Vector multiply(double pScale, Vector pVector) {
+  public static Vector multiply(float pScale, Vector pVector) {
     return new Vector(pScale * pVector.getX(), pScale * pVector.getY(), pScale * pVector.getZ());
   }
 
-  public static Vector inverseMultiply(double pScale, Vector pVector) {
+  public static Vector inverseMultiply(float pScale, Vector pVector) {
     return new Vector(pVector.getX() / pScale, pVector.getY() / pScale, pVector.getZ() / pScale);
   }
 
-  public static double len(Vector pVector) {
-    return MathLib.sqrt(pVector.getX() * pVector.getX() + pVector.getY() * pVector.getY() + pVector.getZ() * pVector.getZ());
+  public static float len(Vector pVector) {
+    return (float) MathLib.sqrt(pVector.getX() * pVector.getX() + pVector.getY() * pVector.getY() + pVector.getZ() * pVector.getZ());
   }
 
   public static Vector normalize(Vector pVector) {
-    double r = len(pVector);
+    float r = len(pVector);
     if (r > MathLib.EPSILON) {
       return inverseMultiply(r, pVector);
     }
@@ -60,8 +60,8 @@ public class VectorMath {
   public static Vector tangent(Vector pA, Vector pCentre) {
     Vector r = normal(pA, pCentre);
 
-    Vector c1 = cross(r, new Vector(0.0, 0.0, 1.0));
-    Vector c2 = cross(r, new Vector(0.0, 1.0, 0.0));
+    Vector c1 = cross(r, new Vector(0.0f, 0.0f, 1.0f));
+    Vector c2 = cross(r, new Vector(0.0f, 1.0f, 0.0f));
 
     if (len(c1) > len(c2)) {
       return c1;
